+++ CHANGELOG.md
@@ -38,6 +38,7 @@ Other improvements:
   improves quality of the flow-matching samplers  
 - t5 support manually downloaded models  
   applies to all models that use t5 transformer  
+- fix handling of model configs if offline config is not available  
 
 Work-in-progress:
 - flux controlnet support: (*1)